requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
azure-core==1.34.0
azure-storage-blob==12.25.1
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cloudpickle==3.1.1
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
-e git+https://github.com/facebookresearch/detectron2.git@536dc9d527074e3b15df5f6677ffe1f4e104a4ab#egg=detectron2
fastapi==0.115.12
filelock==3.13.1
fonttools==4.58.0
fsspec==2024.6.1
fvcore==0.1.5.post20221221
grpcio==1.71.0
h11==0.16.0
hydra-core==1.3.2
idna==3.10
iopath==0.1.9
isodate==0.7.2
Jinja2==3.1.4
joblib==1.5.0
kiwisolver==1.4.8
Markdown==3.8
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.8
portalocker==3.1.1
protobuf==6.30.2
pycocotools==2.0.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
setuptools==70.2.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
torch==2.7.0+cu118
torchaudio==2.7.0+cu118
torchvision==0.22.0+cu118
tqdm==4.67.1
triton==3.3.0
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.4.0
uvicorn==0.34.2
Werkzeug==3.1.3
wheel==0.45.1
yacs==0.1.8
